Fixed home button logic after a player win game
It was not redirecting to your home page after a player win game, as earlier it was: "window.location.href='index.html'" .

It should be ../index.html. Thats why I called, "renderHomePage()" function for better clarity.

But, I must say the project is good, specially the melody matcha.
All the best, for future projects. Hoping to collaborate on a single project in near future :-)